KSENROLL-4879:Added joint defined Co info.
git-svn-id: https://svn.kuali.org/repos/student/enrollment/ks-enroll/trunk@45178 0ac02d80-fb3f-11dc-a574-c57244a850eb

+++ ks-enroll-ui/src/main/java/org/kuali/student/enrollment/class2/courseoffering/controller/CourseOfferingManagementController.java
@@ -860,7 +860,10 @@ public class CourseOfferingManagementController extends UifControllerBase  {
             boolean hasDeletion = true;
             if (co.getIsChecked()) {
                 checked++;
-                CourseOfferingResourceLoader.loadCourseOfferingService().deleteCourseOfferingCascaded(co.getCourseOfferingId(), ContextBuilder.loadContextInfo());
+                if(!co.isColocated()) {
+                    // dol not delete colocated COs
+                    CourseOfferingResourceLoader.loadCourseOfferingService().deleteCourseOfferingCascaded(co.getCourseOfferingId(), ContextBuilder.loadContextInfo());
+                }
             }
         }
